# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Habit Journey API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkins:
    get:
      operationId: listHabitCheckins
      tags: [checkins]
      summary: List check-ins for a habit
      parameters:
        - name: habitId
          in: query
          required: true
          schema: { type: integer }
        - name: from
          in: query
          required: false
          schema: { type: string, format: date }
        - name: to
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "200":
          description: Check-ins
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Checkin" }
  /habits/{habitId}/checkins:
    post:
      operationId: createCheckin
      tags: [checkins]
      summary: Record a check-in for a habit (updates streak and awards coins)
      parameters:
        - name: habitId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckinInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckinResult" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: A successful check-in cannot be changed to incomplete
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /habits/{habitId}/checkins/{date}:
    patch:
      operationId: updateCheckinReflection
      tags: [checkins]
      summary: Add or update reflection and missed-day reason for a recorded check-in
      parameters:
        - name: habitId
          in: path
          required: true
          schema: { type: integer }
        - name: date
          in: path
          required: true
          schema: { type: string, format: date }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckinReflection" }
      responses:
        "200":
          description: Updated reflection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Checkin" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /habits/{habitId}/recover-streak:
    post:
      operationId: recoverStreak
      tags: [checkins]
      summary: Spend coins to restore a just-broken streak
      parameters:
        - name: habitId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Streak recovered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecoverStreakResult" }
        "400":
          description: Not eligible or insufficient coins
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /dashboard/calendar:
    get:
      operationId: getDashboardCalendar
      tags: [dashboard]
      summary: Daily completion heatmap data for a calendar month
      parameters:
        - name: month
          in: query
          required: true
          description: "Format YYYY-MM"
          schema: { type: string }
      responses:
        "200":
          description: Calendar days
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DashboardCalendarDay" }
...
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    HabitMilestone:
      type: object
      required: [title, description, targetValue, order]
      properties:
        title: { type: string }
        description: { type: string }
        targetValue: { type: number }
        order: { type: integer }
    Habit:
      type: object
      required: [id, title, emoji, category, cadence, customDays, unit, targetValue, minimumValue, busyDayValue, baselineValue, successLimitValue, cueType, cueTime, cue, startAction, friction, minimumFloor, journeyStartDate, journeyLength, rewardId, difficulty, goalType, isActive, currentStreak, longestStreak, lastCheckinDate, milestones, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        emoji: { type: string }
        category: { type: string, enum: [health, learning, productivity, mindfulness, social, creativity, finance, custom] }
        cadence: { type: string, enum: [daily, weekdays, weekly, custom_days] }
        customDays:
          type: ["array", "null"]
          items: { type: integer, minimum: 0, maximum: 6 }
        unit: { type: string, enum: [minutes, count, pages, custom] }
        executionType: { type: ["string", "null"], enum: [duration, count, boolean, limit, null], description: Nullable for legacy records; inferred from unit and goalType when consumed. }
        targetValue: { type: number }
        minimumValue: { type: ["number", "null"] }
        busyDayValue: { type: ["number", "null"] }
        baselineValue: { type: ["number", "null"] }
        successLimitValue: { type: ["number", "null"] }
        cueType: { type: ["string", "null"], enum: [time, routine, custom, null] }
        cueTime: { type: ["string", "null"], pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$' }
        cue: { type: ["string", "null"] }
        startAction: { type: ["string", "null"] }
        friction: { type: ["string", "null"] }
        minimumFloor: { type: ["number", "null"], minimum: 0 }
        journeyStartDate: { type: ["string", "null"], format: date }
        journeyLength: { type: ["integer", "null"], minimum: 1, maximum: 22 }
        journeyCompletedAt:
          type: ["string", "null"]
          format: date-time
          description: Server-recorded journey completion instant; optional for compatibility with older servers.
        rewardId: { type: ["integer", "null"], minimum: 1 }
        difficulty: { type: string, enum: [easy, medium, hard] }
        goalType: { type: string, enum: [build, quit] }
        isActive: { type: boolean }
        currentStreak: { type: integer }
        longestStreak: { type: integer }
        lastCheckinDate: { type: ["string", "null"], format: date }
        milestones:
          type: array
          items: { $ref: "#/components/schemas/HabitMilestone" }
        createdAt: { type: string, format: date-time }
...
    Checkin:
      type: object
      required: [id, habitId, date, completed, value, note, moodRating, difficulty, missedReason, targetSnapshot, minimumSnapshot, successLimitSnapshot, goalTypeSnapshot, targetCompleted, coinsEarned, xpEarned, createdAt]
      properties:
        id: { type: integer }
        habitId: { type: integer }
        date: { type: string, format: date }
        completed: { type: boolean }
        value: { type: ["number", "null"] }
        note: { type: ["string", "null"] }
        moodRating: { type: ["integer", "null"] }
        difficulty: { type: ["string", "null"], enum: [easy, normal, hard, very_hard, null] }
        missedReason: { type: ["string", "null"], enum: [too_difficult, no_time, forgot, lost_motivation, unexpected, other, null] }
        targetSnapshot: { type: ["number", "null"] }
        minimumSnapshot: { type: ["number", "null"] }
        successLimitSnapshot: { type: ["number", "null"] }
        goalTypeSnapshot: { type: ["string", "null"], enum: [build, quit, null] }
        targetCompleted: { type: boolean }
        coinsEarned: { type: integer }
        xpEarned: { type: ["integer", "null"], description: Exact XP persisted atomically with newly granted check-in rewards; null means historical earnings are unknown }
        createdAt: { type: string, format: date-time }
    CheckinInput:
      type: object
      required: [date]
      properties:
        date: { type: string, format: date }
        completed: { type: boolean, description: Legacy compatibility; success is determined server-side when value is provided }
        value: { type: number }
        note: { type: string }
        moodRating: { type: integer, minimum: 1, maximum: 5 }
        difficulty: { type: string, enum: [easy, normal, hard, very_hard] }
        missedReason: { type: string, enum: [too_difficult, no_time, forgot, lost_motivation, unexpected, other] }
    CheckinReflection:
      type: object
      properties:
        note: { type: ["string", "null"] }
        moodRating: { type: ["integer", "null"], minimum: 1, maximum: 5 }
        difficulty: { type: ["string", "null"], enum: [easy, normal, hard, very_hard, null] }
        missedReason: { type: ["string", "null"], enum: [too_difficult, no_time, forgot, lost_motivation, unexpected, other, null] }
...
    CheckinResult:
      allOf:
        - $ref: "#/components/schemas/Checkin"
        - type: object
          required: [newStreak, habit]
          properties:
            newStreak: { type: integer }
            habit: { $ref: "#/components/schemas/Habit" }
...
    RecoverStreakResult:
      type: object
      required: [habit, coinsSpent]
      properties:
        habit: { $ref: "#/components/schemas/Habit" }
        coinsSpent: { type: integer }
...
    DashboardCalendarDay:
      type: object
      required: [date, completedCount, scheduledCount, timeMinutes]
      properties:
        date: { type: string, format: date }
        completedCount: { type: integer }
        scheduledCount: { type: integer }
        timeMinutes: { type: integer }
...