# lib/api-zod/src/generated/types/habit.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Habit Journey API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HabitCadence } from './habitCadence';
import type { HabitCategory } from './habitCategory';
import type { HabitCueType } from './habitCueType';
import type { HabitDifficulty } from './habitDifficulty';
import type { HabitExecutionType } from './habitExecutionType';
import type { HabitGoalType } from './habitGoalType';
import type { HabitMilestone } from './habitMilestone';
import type { HabitUnit } from './habitUnit';

export interface Habit {
  id: number;
  title: string;
  emoji: string;
  category: HabitCategory;
  cadence: HabitCadence;
  /**
     * @nullable
     * @items.minimum 0
     * @items.maximum 6
     */
  customDays: number[] | null;
  unit: HabitUnit;
  /**
     * Nullable for legacy records; inferred from unit and goalType when consumed.
     * @nullable
     */
  executionType?: HabitExecutionType;
  targetValue: number;
  /** @nullable */
  minimumValue: number | null;
  /** @nullable */
  busyDayValue: number | null;
  /** @nullable */
  baselineValue: number | null;
  /** @nullable */
  successLimitValue: number | null;
  /** @nullable */
  cueType: HabitCueType;
  /**
     * @nullable
     * @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$
     */
  cueTime: string | null;
  /** @nullable */
  cue: string | null;
  /** @nullable */
  startAction: string | null;
  /** @nullable */
  friction: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  minimumFloor: number | null;
  /** @nullable */
  journeyStartDate: Date | null;
  /**
     * @minimum 1
     * @maximum 22
     * @nullable
     */
  journeyLength: number | null;
  /**
     * Server-recorded journey completion instant; optional for compatibility with older servers.
     * @nullable
     */
  journeyCompletedAt?: Date | null;
  /**
     * @minimum 1
     * @nullable
     */
  rewardId: number | null;
  difficulty: HabitDifficulty;
  goalType: HabitGoalType;
  isActive: boolean;
  currentStreak: number;
  longestStreak: number;
  /** @nullable */
  lastCheckinDate: Date | null;
  milestones: HabitMilestone[];
  createdAt: Date;
}